Change to decimal setup

+++ lib/otr_bunq/donation.ex
@@ -2,7 +2,7 @@ defmodule OtrBunq.Donation do
   use Ecto.Schema
 
   schema "donations" do
-    field(:amount, :float)
+    field(:amount, :decimal)
     field(:timestamp, :utc_datetime)
     field(:bunq_payment_id, :integer)
     field(:description, :string)
